rocm/gfx1151: F16 mid output, always-cache IQ2 LUTs, skip sorting for small batches
P0.1 — F16 mid output for MoE
  All 27 moe_gate_up_mid_* kernels write __half mid instead of float.
  Down kernels read F16 via MID_F16=true or direct __half* parameters.
  Added f16_to_f32_kernel for non-batched fallback.
  mid_is_f16=true, allocation halved to sizeof(uint16_t).

P0.2 — Always cache IQ2_XXS LUTs in shared memory
  Removed xq_blocks <= 16u conditional from LUT loading in 5 IQ2 kernels.
  LUTs (2176 B) always cached; Q8_K block cache still respects capacity.
  use_decode_lut_gate always true for decode.

P0.3 — Reduce sorted-pair overhead for small batches
  Changed threshold from n_tokens > 1 to n_tokens > 8.
  Batch sizes 2-8 skip 3 sorting kernels (count, prefix, scatter).
  Saves ~50-100 us per layer for small-batch decode.

Benchmarked on IQ2_XXS-w2-Q2K model (5 runs each):
  Baseline: prefill 51.20 t/s, gen 15.67 t/s
  All P0:   prefill 55.11 t/s (+7.6%), gen 15.68 t/s (+0.1%)

+++ ds4.c
@@ -11211,7 +11211,7 @@ static bool metal_graph_alloc_raw_cap(
                              sizeof(int32_t));
     g->batch_routed_gate = ds4_gpu_tensor_alloc(pc * DS4_N_EXPERT_USED * routed_mid_dim * sizeof(float));
     g->batch_routed_up = ds4_gpu_tensor_alloc(pc * DS4_N_EXPERT_USED * routed_mid_dim * sizeof(float));
-    g->batch_routed_mid = ds4_gpu_tensor_alloc(pc * DS4_N_EXPERT_USED * routed_mid_dim * sizeof(float));
+    g->batch_routed_mid = ds4_gpu_tensor_alloc(pc * DS4_N_EXPERT_USED * routed_mid_dim * sizeof(uint16_t));
     g->batch_routed_down = ds4_gpu_tensor_alloc(pc * DS4_N_EXPERT_USED * DS4_N_EMBD * sizeof(float));
     g->batch_routed_out = ds4_gpu_tensor_alloc(pc * DS4_N_EMBD * sizeof(float));
